requirements.txt
moviepy
imageio
pillow
pillow
opencv-python
pytesseract
python-dotenv
requests
pydantic
tqdm